createArea refactored
